+++ packages/fs-detectors/test/unit.examples.test.ts
@@ -29,21 +29,21 @@ const experimentalExamples = new Set([
 ]);
 
 describe('examples should be detected', () => {
-  it.each(getExamples())(
-    'should detect $exampleName',
-    async ({ exampleName, examplePath }) => {
-      const fs = new LocalFileSystemDetector(examplePath);
-      const useExperimentalFrameworks = experimentalExamples.has(exampleName);
-      const framework = await detectFramework({
-        fs,
-        frameworkList,
-        useExperimentalFrameworks,
-      });
-      if (!framework) {
-        throw new Error(`Framework not detected for example "${exampleName}".`);
-      }
-
-      expect(framework).toBe(overrides.get(framework) ?? framework);
+  it.each(getExamples())('should detect $exampleName', async ({
+    exampleName,
+    examplePath,
+  }) => {
+    const fs = new LocalFileSystemDetector(examplePath);
+    const useExperimentalFrameworks = experimentalExamples.has(exampleName);
+    const framework = await detectFramework({
+      fs,
+      frameworkList,
+      useExperimentalFrameworks,
+    });
+    if (!framework) {
+      throw new Error(`Framework not detected for example "${exampleName}".`);
     }
-  );
+
+    expect(framework).toBe(overrides.get(framework) ?? framework);
+  });
 });