+++ packages/cervel/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vercel/cervel
 
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/backends@0.0.40
+
+## 0.0.26
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/backends@0.0.39
+
 ## 0.0.25
 
 ### Patch Changes