+++ packages/express/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @vercel/express
 
+## 0.1.52
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.6.10
+  - @vercel/cervel@0.0.27
+
+## 0.1.51
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.6.9
+  - @vercel/cervel@0.0.26
+
 ## 0.1.50
 
 ### Patch Changes