+++ packages/static-build/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vercel/static-build
 
+## 2.8.44
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/gatsby-plugin-vercel-builder@2.0.142
+
+## 2.8.43
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/gatsby-plugin-vercel-builder@2.0.141
+
 ## 2.8.42
 
 ### Patch Changes